Reindent to clear warnings from build process
According to the following issue:
https://github.com/openebs/openebs/issues/400

There were a lot of warnings on `make html` that ought to be resolved
I went through the steps of patching the relevant file until I no
longer got any build warnings myself and these are the results.

However, these changes introduce awkward formatting in the source
files so I'm not necessarily sure this is the best way to go.